+++ apps/node/src/index.ts
@@ -1,14 +1,40 @@
 import DeviceConnection from '@cypherock/sdk-hw-hid';
+import DeviceConnectionSerialport from '@cypherock/sdk-hw-serialport';
 import { ManagerApp } from '@cypherock/sdk-app-manager';
+import { IDeviceConnection } from '@cypherock/sdk-interfaces';
 
 const run = async () => {
-  const connection = await DeviceConnection.create();
+  let connection: IDeviceConnection;
+
+  try {
+    connection = await DeviceConnection.create();
+  } catch (error) {
+    connection = await DeviceConnectionSerialport.create();
+  }
 
   const managerApp = await ManagerApp.create(connection);
 
   const deviceInfo = await managerApp.getDeviceInfo();
 
   console.log(deviceInfo);
+
+  await managerApp.authDevice();
+
+  await managerApp.trainCard({ onWallets: async () => true });
+
+  await managerApp.authCard();
+
+  // await managerApp.updateFirmware({
+  //   getDevices: async () => [
+  //     ...(await DeviceConnection.list()),
+  //     ...(await DeviceConnectionSerialport.list()),
+  //   ],
+  //   createConnection: async d =>
+  //     d.type === 'hid'
+  //       ? DeviceConnection.connect(d)
+  //       : DeviceConnectionSerialport.connect(d),
+  //   allowPrerelease: true,
+  // });
 };
 
 run();