+++ packages/app-manager/src/app.ts
@@ -2,6 +2,8 @@ import { IDeviceConnection } from '@cypherock/sdk-interfaces';
 import { SDK } from '@cypherock/sdk-core';
 
 import * as operations from './operations';
+import { firmwareService } from './services';
+import { GetLatestFirmwareOptions } from './services/firmware';
 
 export class ManagerApp {
   private readonly sdk: SDK;
@@ -53,11 +55,21 @@ export class ManagerApp {
     );
   }
 
+  public async updateFirmware(params: operations.IUpdateFirmwareParams) {
+    return this.sdk.runOperation(() =>
+      operations.updateFirmware(this.sdk, params),
+    );
+  }
+
   public async trainCard(params: operations.ITrainCardParams) {
     return this.sdk.runOperation(() => operations.trainCard(this.sdk, params));
   }
 
   public async destroy() {
     return this.sdk.destroy();
+  }
+
+  public static async getLatestFirmware(params?: GetLatestFirmwareOptions) {
+    return firmwareService.getLatest(params);
   }
 }