+++ packages/app-manager/src/proto/types.ts
@@ -1,0 +1,7 @@
+export enum UpdateFirmwareStatus {
+  UPDATE_FIRMWARE_STATUS_INIT = 0,
+  UPDATE_FIRMWARE_STATUS_USER_CONFIRMED = 1,
+  UNRECOGNIZED = -1,
+}
+
+export * from './generated/types';