+++ packages/app-manager/tests/02.getDeviceInfo/__fixtures__/error.ts
@@ -14,12 +14,11 @@ const withUnknownError: IGetDeviceInfoTestCase = {
 };
 
 const withInvalidAppId: IGetDeviceInfoTestCase = {
-  name: 'With msg from device',
+  name: 'With corrupt msg from device',
   query: new Uint8Array([10, 2, 10, 0]),
   result: new Uint8Array([10, 4, 18, 2, 16, 0]),
   errorInstance: DeviceAppError,
-  errorMessage:
-    deviceAppErrorTypeDetails[DeviceAppErrorType.INVALID_MSG_FROM_DEVICE],
+  errorMessage: deviceAppErrorTypeDetails[DeviceAppErrorType.CORRUPT_DATA],
 };
 
 const error: IGetDeviceInfoTestCase[] = [withUnknownError, withInvalidAppId];