+++ packages/util/src/utils/index.ts
@@ -4,3 +4,5 @@ export * from './sleep';
 export * from './createStatusListener';
 export * from './logger';
 export * from './config';
+export * from './version';
+export * from './querystring';